backend/requirements.txt
fastapi
uvicorn
mysqlclient
SQLAlchemy
pymysql
python-dotenv
python-jose
python-multipart
flask
flask-cors
pandas